Style changes from review:
- Moved size calculation to be part of BlockHeader for better encapsulation.
- Changed internal size calculations to use usize for consistency with typical `.size()` methods, converting to u64 only at the point of returning from the API.
- Refactored size functions to use a single big sum instead of starting with `let mut size` and then incrementing with `+=`. This should compile to the same thing but appears cleaner.

+++ zilliqa/src/api/types/eth.rs
@@ -108,7 +108,7 @@ impl Block {
     pub fn from_block(block: &message::Block, miner: Address, block_gas_limit: EvmGas) -> Self {
         Block {
             header: Header::from_header(block.header, miner, block_gas_limit),
-            size: block.size(),
+            size: block.size() as u64,
             transactions: block
                 .transactions
                 .iter()
